Clean up naming of views endpoints in api browser (#7178)
    * Clean up naming of views-related resources in api browser

    - removed "Enterprise/" prefix
    - removed deprecated (and unused) property "description"

    * Added prefix "Legacy/" to old search resources in api browser

    * Changed name of new SavedSearchesResource in api browser

    Changed "SavedSearches" to "Search/Saved" to consistently align it with other search endpoints

    * Fixed `ApiOperation` description in `DashboardsResource`

    * Added prefix "Legacy/" to old dashboard resources in api browser

    * Changed api browser name of `ViewSharingResource`...

    from `Views` to `ViewSharing`

+++ dataset/GitHub_Java/Graylog2.graylog2-server/graylog2-server/src/main/java/org/graylog/plugins/views/search/rest/ViewSharingResource.java
@@ -21,7 +21,6 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
 import org.graylog.plugins.views.audit.ViewsAuditEventTypes;
-import org.graylog.plugins.views.search.views.ViewDTO;
 import org.graylog.plugins.views.search.views.ViewService;
 import org.graylog.plugins.views.search.views.sharing.UserShortSummary;
 import org.graylog.plugins.views.search.views.sharing.ViewSharing;
@@ -30,13 +29,10 @@ import org.graylog2.audit.jersey.AuditEvent;
 import org.graylog2.plugin.database.users.User;
 import org.graylog2.plugin.rest.PluginRestResource;
 import org.graylog2.shared.rest.resources.RestResource;
-import org.graylog2.shared.security.RestPermissions;
 import org.graylog2.shared.users.UserService;
 
 import javax.inject.Inject;
-import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.NotFoundException;
@@ -75,9 +71,9 @@ public class ViewSharingResource extends RestResource implements PluginRestResou
     @POST
     @ApiOperation("Configure sharing for a view")
     @AuditEvent(type = ViewsAuditEventTypes.VIEW_SHARING_CREATE)
-    public ViewSharing create(@ApiParam(name="id") @PathParam("id") @NotEmpty String id, @Valid @NotNull ViewSharing viewSharing) {
+    public ViewSharing create(@ApiParam(name="id") @PathParam("id") @NotEmpty String id, ViewSharing viewSharing) {
         ensureUserIsPermittedForView(id);
-        ensureUserIsPermittedToEditView(id);
+        checkPermission(ViewsRestPermissions.VIEW_EDIT, id);
         return viewSharingService.create(viewSharing);
     }
 
@@ -86,7 +82,7 @@ public class ViewSharingResource extends RestResource implements PluginRestResou
     @AuditEvent(type = ViewsAuditEventTypes.VIEW_SHARING_DELETE)
     public ViewSharing delete(@ApiParam(name="id") @PathParam("id") @NotEmpty String id) {
         ensureUserIsPermittedForView(id);
-        ensureUserIsPermittedToEditView(id);
+        checkPermission(ViewsRestPermissions.VIEW_EDIT, id);
         return viewSharingService.remove(id).orElse(null);
     }
 
@@ -103,26 +99,7 @@ public class ViewSharingResource extends RestResource implements PluginRestResou
     }
 
     private void ensureUserIsPermittedForView(String viewId) {
-        final ViewDTO view = viewService.get(viewId).orElseThrow(NotFoundException::new);
-        if (view.type().equals(ViewDTO.Type.DASHBOARD)) {
-            checkAnyPermission(new String[]{
-                    RestPermissions.DASHBOARDS_READ,
-                    ViewsRestPermissions.VIEW_READ
-            }, viewId);
-        } else {
-            checkPermission(ViewsRestPermissions.VIEW_READ, viewId);
-        }
-    }
-
-    private void ensureUserIsPermittedToEditView(String viewId) {
-        final ViewDTO view = viewService.get(viewId).orElseThrow(NotFoundException::new);
-        if (view.type().equals(ViewDTO.Type.DASHBOARD)) {
-            checkAnyPermission(new String[]{
-                    RestPermissions.DASHBOARDS_EDIT,
-                    ViewsRestPermissions.VIEW_EDIT
-            }, viewId);
-        } else {
-            checkPermission(ViewsRestPermissions.VIEW_EDIT, viewId);
-        }
+        viewService.get(viewId).orElseThrow(NotFoundException::new);
+        checkPermission(ViewsRestPermissions.VIEW_READ, viewId);
     }
 }
